for #2556, revert ShardingInsertOptimizer
