+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/filler/sharding/dml/insert/ShardingInsertValuesFiller.java
@@ -23,13 +23,19 @@ import org.apache.shardingsphere.core.metadata.table.ShardingTableMetaData;
 import org.apache.shardingsphere.core.parse.aware.ShardingRuleAware;
 import org.apache.shardingsphere.core.parse.aware.ShardingTableMetaDataAware;
 import org.apache.shardingsphere.core.parse.filler.SQLSegmentFiller;
+import org.apache.shardingsphere.core.parse.sql.context.condition.AndCondition;
+import org.apache.shardingsphere.core.parse.sql.context.condition.Column;
+import org.apache.shardingsphere.core.parse.sql.context.condition.Condition;
 import org.apache.shardingsphere.core.parse.sql.context.insertvalue.InsertValue;
 import org.apache.shardingsphere.core.parse.sql.segment.dml.InsertValuesSegment;
+import org.apache.shardingsphere.core.parse.sql.segment.dml.expr.ExpressionSegment;
+import org.apache.shardingsphere.core.parse.sql.segment.dml.expr.simple.SimpleExpressionSegment;
 import org.apache.shardingsphere.core.parse.sql.statement.SQLStatement;
 import org.apache.shardingsphere.core.parse.sql.statement.dml.InsertStatement;
 import org.apache.shardingsphere.core.rule.ShardingRule;
 
 import java.util.Collection;
+import java.util.Iterator;
 
 /**
  * Insert values filler for sharding.
@@ -49,6 +55,7 @@ public final class ShardingInsertValuesFiller implements SQLSegmentFiller<Insert
         InsertStatement insertStatement = (InsertStatement) sqlStatement;
         fillColumns(sqlSegment, insertStatement);
         fillValues(sqlSegment, insertStatement);
+        fillCondition(sqlSegment, insertStatement);
     }
     
     private void fillColumns(final InsertValuesSegment sqlSegment, final InsertStatement insertStatement) {
@@ -76,5 +83,22 @@ public final class ShardingInsertValuesFiller implements SQLSegmentFiller<Insert
     
     private void fillValues(final InsertValuesSegment sqlSegment, final InsertStatement insertStatement) {
         insertStatement.getValues().add(new InsertValue(sqlSegment.getValues()));
+    }
+    
+    private void fillCondition(final InsertValuesSegment segment, final InsertStatement insertStatement) {
+        AndCondition andCondition = new AndCondition();
+        Iterator<String> columnNames = insertStatement.getColumnNames().iterator();
+        for (ExpressionSegment each : segment.getValues()) {
+            if (each instanceof SimpleExpressionSegment) {
+                fillShardingCondition(andCondition, insertStatement.getTables().getSingleTableName(), columnNames.next(), (SimpleExpressionSegment) each);
+            }
+        }
+        insertStatement.getShardingConditions().getOrConditions().add(andCondition);
+    }
+    
+    private void fillShardingCondition(final AndCondition andCondition, final String tableName, final String columnName, final SimpleExpressionSegment expressionSegment) {
+        if (shardingRule.isShardingColumn(columnName, tableName)) {
+            andCondition.getConditions().add(new Condition(new Column(columnName, tableName), null, expressionSegment));
+        }
     }
 }
