+++ sharding-core/sharding-core-parse/sharding-core-parse-sqlserver/src/main/resources/META-INF/parsing-rule-definition/sharding/sqlserver/sql-statement-rule-definition.xml
@@ -18,7 +18,7 @@
 
 <sql-statement-rule-definition>
     <sql-statement-rule context="select" sql-statement-class="org.apache.shardingsphere.core.parse.sql.statement.dml.SelectStatement" extractor-rule-refs="tableReferences, columns, selectItems, subqueryPredicate, where, predicate, groupBy, orderBy" optimizer-class="org.apache.shardingsphere.core.parse.optimizer.select.ShardingSelectOptimizer" />
-    <sql-statement-rule context="insert" sql-statement-class="org.apache.shardingsphere.core.parse.sql.statement.dml.InsertStatement" extractor-rule-refs="table, columns, insertColumns, insertValues" optimizer-class="org.apache.shardingsphere.core.parse.optimizer.insert.ShardingInsertOptimizer" />
+    <sql-statement-rule context="insert" sql-statement-class="org.apache.shardingsphere.core.parse.sql.statement.dml.InsertStatement" extractor-rule-refs="table, columns, insertColumns, insertValues" />
     <sql-statement-rule context="update" sql-statement-class="org.apache.shardingsphere.core.parse.sql.statement.dml.UpdateStatement" extractor-rule-refs="tableReferences, columns, setAssignments, where, predicate" />
     <sql-statement-rule context="delete" sql-statement-class="org.apache.shardingsphere.core.parse.sql.statement.dml.DeleteStatement" extractor-rule-refs="tables, columns, where, predicate" />
